Calculator class and sum method
We want to be able to follow the test - code - refactor process to
do TDD.

This implements the tests for math methods first and the the code
for this to actually work and perform the sum action.

And at the end the refactor process to make the code look better
and more mantainable.